Add S3 distribution upload functionality and update configuration variables
