+++ .github/workflows/deploy-sdk-production.yaml
@@ -41,7 +41,8 @@ jobs:
       CONFIG_URL_RAYS: ${{ secrets.CONFIG_URL_RAYS }}
       SUMMER_HUB_CHAIN_ID: ${{ secrets.SDK_HUB_CHAIN_ID }}
       SUMMER_DEPLOYED_CHAINS_ID: ${{ secrets.SUMMER_DEPLOYED_CHAINS_ID }}
-      SDK_DISTRIBUTIONS_URLS: ${{ secrets.SDK_DISTRIBUTIONS_URLS }}
+      SDK_DISTRIBUTIONS_BASE_URL: ${{ secrets.SDK_DISTRIBUTIONS_BASE_URL }}
+      SDK_DISTRIBUTIONS_FILES: ${{ secrets.SDK_DISTRIBUTIONS_FILES }}
     steps:
       - name: Send start notification to Discord
         uses: Ilshidur/action-discord@0.3.2
