+++ package.json
@@ -40,6 +40,7 @@
     "armada:upconfig": "cd armada-protocol/contracts/packages/deployment/config && cp -r index.json ../../../../../sdk/armada-protocol-common/src/deployments/config.json"
   },
   "devDependencies": {
+    "@aws-sdk/client-s3": "^3.732.0",
     "@graphql-codegen/add": "^5.0.2",
     "@graphql-codegen/cli": "^5.0.2",
     "@graphql-codegen/typed-document-node": "^5.0.6",
@@ -81,7 +82,6 @@
   },
   "dependencies": {
     "bignumber.js": "^9.1.2",
-    "kysely": "^0.27.3",
     "viem": "2.21.55",
     "zod": "^3.22.4"
   },
