+++ stacks/package.json
@@ -8,16 +8,16 @@
     "lint": "eslint .",
     "lint:fix": "eslint . --fix"
   },
+  "type": "module",
   "devDependencies": {
     "@summerfi/eslint-config": "workspace:*",
-    "@summerfi/typescript-config": "workspace:*",
     "@summerfi/rays-db": "workspace:*",
+    "@summerfi/typescript-config": "workspace:*",
+    "@types/pg": "^8.11.5",
     "eslint": "^8.56.0",
     "kysely": "^0.27.3",
-    "pg": "^8.11.5",
-    "@types/pg": "^8.11.5",
-    "tsx": "^4.8.2",
-    "kysely-codegen": "^0.15.0"
+    "kysely-codegen": "^0.15.0",
+    "pg": "^8.11.5"
   },
   "private": true
 }
