+++ turbo.json
@@ -82,7 +82,8 @@
     "SDK_API_URL",
     "SDK_USE_FORK",
     "SDK_FORK_CONFIG",
-    "SDK_DISTRIBUTIONS_URLS",
+    "SDK_DISTRIBUTIONS_BASE_URL",
+    "SDK_DISTRIBUTIONS_FILES",
     "EARN_PROTOCOL_DB_CONNECTION_STRING",
     "EARN_PROTOCOL_JWT_SECRET",
     "EARN_PROTOCOL_JWT_CHALLENGE_SECRET",
